refactor: moveアクション処理を関数化
